Switch to Installation::Widgets::Overview.new(client: "foo_proposal")
they turned out to be similar enough to work as instances of the same class
