+++ src/Makefile.am
@@ -215,6 +215,7 @@ ylibdialog_DATA = \
 ylibwidgetdir = "${yast2dir}/lib/installation/widgets"
 ylibwidget_DATA = \
   lib/installation/widgets/mocked.rb \
+  lib/installation/widgets/overview.rb \
   lib/installation/widgets/system_role.rb
 
 EXTRA_DIST = $(module_DATA) $(client_DATA) $(ynclude_DATA) $(scrconf_DATA) $(schemafiles_DATA) $(desktop_DATA) $(fillup_DATA) $(ylibdialog_DATA) $(ylib_DATA) $(ylibtransfer_DATA)
